+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [2.11.1](https://github.com/alessandroraffa/tangyr-vscode/compare/v2.11.0...v2.11.1) (2026-09-04)
+
+### Bug Fixes
+
+* **ci:** pin compatible vscode test host ([bf29269](https://github.com/alessandroraffa/tangyr-vscode/commit/bf292695823709caa50e981b86f6461b3cbdb33e))
+
+## [2.11.0](https://github.com/alessandroraffa/tangyr-vscode/compare/v2.10.1...v2.11.0) (2026-09-04)
+
+### Features
+
+* **archiving:** import explicit session sources ([e4f0b90](https://github.com/alessandroraffa/tangyr-vscode/commit/e4f0b909123d0ec851709f28b25d0925424328b0))
+
 ## [2.10.1](https://github.com/alessandroraffa/tangyr-vscode/compare/v2.10.0...v2.10.1) (2026-07-12)
 
 ### Bug Fixes